dom/tests/mochitest/general/test_bug1279771.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1279771
-->
<head>
  <title>Test for Bug 1279771</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script src="/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <select id="select">
    <option>A</option>
    <option>B</option>
    <option>C</option>
  </select>

<pre id="test">
  <script type="application/javascript">
    add_task(function* (){
      var select = document.querySelector("#select");
      let done = new Promise(resolve => {
        select.addEventListener('change',  function() {
          ok(document.execCommand("copy"), "The copy command is allowed to be executed from this callback");
          resolve();
        });
      })

      // Spin the event loop once so that synthesizeMouseAtCenter works...
      yield new Promise(resolve => SimpleTest.executeSoon(resolve));
      synthesizeMouseAtCenter(select, {});

      // Run the chrome script which will close the selectmenulist, firing a change event
      SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL("file_bug1279771_chromeScript.js"));
      yield done;
    });
  </script>
</pre>
</body>
</html>
